saved changes
Просто сохранил изменения с устройства Dell

+++ My2DGame/Assets/ECS/ECSMonoBehaviour Scripts/MB Enemy View/EnemyView.cs
@@ -6,7 +6,7 @@ public class EnemyView : MonoBehaviour
     public float meleeAttackDistance;
     public float triggerDistance;
     public float meleeAttackInterval;
-    public int startHealth;
-    public int damage;
+    public float startHealth;
+    public float damage;
     public EcsEntity entity;
 }
